algorithm/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
geopy==2.2.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0